fix imprecise statement

+++ src/pynxtools_apm/utils/oasis_apsuite_reader.py
@@ -69,9 +69,9 @@ class NxApmNomadOasisCamecaParser:
         trg = f"/ENTRY[entry{self.entry_id}]/atom_probe/hit_finding/"
         for cameca_name, nexus_name in event_type_names:
             if f"fTotalEvent{cameca_name}" in self.yml.as_dict().keys():
-                template[f"{trg}total_event_{nexus_name}"] = self.yml[
-                    f"fTotalEvent{cameca_name}"
-                ]
+                template[f"{trg}total_event_{nexus_name}"] = np.uint64(
+                    self.yml[f"fTotalEvent{cameca_name}"]
+                )
         return template
 
     def parse_versions(self, template: dict) -> dict:
